moving getVoteSaltHash from dev-utils to utils package. Removing listeners file since we can just add listeners in the script for each page
