+++ packages/dev-utils/package.json
@@ -25,7 +25,6 @@
     "chai-bignumber": "^2.0.2",
     "debug": "^3.1.0",
     "dirty-chai": "^2.0.1",
-    "ethereumjs-abi": "^0.6.5",
     "glob": "^7.1.2",
     "handlebars": "^4.0.11",
     "mkdirp": "^0.5.1",
