+++ packages/utils/package.json
@@ -22,6 +22,7 @@
     "typescript": "^2.7.2"
   },
   "dependencies": {
+    "ethereumjs-abi": "^0.6.5",
     "rxjs": "^5.5.6"
   },
   "bugs": {
